requirements_notgan.txt
absl-py==0.12.0
astor==0.8.1
astunparse==1.6.3
backcall==0.2.0
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
copulas==0.5.0
ctgan==0.4.2
cycler==0.10.0
decorator==4.4.2
deepecho==0.2.0
Faker==4.14.2
flatbuffers==1.12
gast==0.3.3
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
graphviz==0.16
grpcio==1.32.0
h5py==2.10.0
idna==2.10
imbalanced-learn==0.8.0
importlib-metadata==4.0.1
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.18.0
joblib==1.0.1
jupyter-client==6.1.12
jupyter-core==4.7.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
llvmlite==0.36.0
Markdown==3.3.4
matplotlib==3.4.1
msgpack==1.0.2
msgpack-numpy==0.4.7.1
networkx==2.5.1
numba==0.53.1
numpy==1.20.2
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.9
pandas==1.1.4
parso==0.8.2
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pomegranate==0.14.1
prompt-toolkit==3.0.18
protobuf==3.15.8
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.8.1
pyparsing==2.4.7
pyspark==2.4.5
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.3
rdt==0.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scikit-plot==0.3.7
scipy==1.6.3
sdmetrics==0.3.0
sdv==0.9.1
six==1.15.0
sklearn==0.0
sktime==0.5.3
statsmodels==0.12.2
tabulate==0.8.9
tensorboard==2.5.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
tensorpack==0.9.4
termcolor==1.1.0
text-unidecode==1.3
threadpoolctl==2.1.0
torch==1.8.1
torchvision==0.9.1
tornado==6.1
tqdm==4.60.0
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.4
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
xgboost==1.4.1
zipp==3.4.1
